Fix flakey tracing publicly API by asserting on the number of spans processed by the exporter
